+++ src/getStepStartStates.ts
@@ -77,6 +77,7 @@ import {
   STEP_RM_BATCH_ACCOUNT,
   STEP_RM_BATCH_POOL,
   STEP_RM_BATCH_APPLICATION,
+  STEP_RM_BATCH_CERTIFICATE,
 } from './steps/resource-manager/batch';
 import { hasSubscriptionId } from '.';
 
@@ -137,5 +138,6 @@ export default function getStepStartStates(
     [STEP_RM_BATCH_ACCOUNT]: resourceManager,
     [STEP_RM_BATCH_POOL]: resourceManager,
     [STEP_RM_BATCH_APPLICATION]: resourceManager,
+    [STEP_RM_BATCH_CERTIFICATE]: resourceManager,
   };
 }