+++ src/steps/resource-manager/batch/constants.ts
@@ -7,6 +7,7 @@ import { createResourceGroupResourceRelationshipMetadata } from '../utils/create
 export const STEP_RM_BATCH_ACCOUNT = 'rm-batch-account';
 export const STEP_RM_BATCH_POOL = 'rm-batch-pool';
 export const STEP_RM_BATCH_APPLICATION = 'rm-batch-application';
+export const STEP_RM_BATCH_CERTIFICATE = 'rm-batch-certificate';
 
 export const BatchEntities = {
   BATCH_ACCOUNT: {
@@ -23,6 +24,11 @@ export const BatchEntities = {
     _type: 'azure_batch_application',
     _class: ['Process'],
     resourceName: '[RM] Batch Application',
+  },
+  BATCH_CERTIFICATE: {
+    _type: 'azure_batch_certificate',
+    _class: ['Certificate'],
+    resourceName: '[RM] Batch Certificate',
   },
 };
 
@@ -51,5 +57,16 @@ export const BatchAccountRelationships = {
     sourceType: BatchEntities.BATCH_ACCOUNT._type,
     _class: RelationshipClass.HAS,
     targetType: BatchEntities.BATCH_APPLICATION._type,
+  },
+
+  BATCH_ACCOUNT_HAS_BATCH_CERTIFICATE: {
+    _type: generateRelationshipType(
+      RelationshipClass.HAS,
+      BatchEntities.BATCH_ACCOUNT,
+      BatchEntities.BATCH_CERTIFICATE,
+    ),
+    sourceType: BatchEntities.BATCH_ACCOUNT._type,
+    _class: RelationshipClass.HAS,
+    targetType: BatchEntities.BATCH_CERTIFICATE._type,
   },
 };