+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to
 - Added `azure_batch_account|has|azure_batch_pool` relationships
 - Added `azure_batch_application` entities
 - Added `azure_batch_account|has|azure_batch_application` relationships
+- Added `azure_batch_certificate` entities
+- Added `azure_batch_account|has|azure_batch_certificate` relationships
 
 ## 5.3.1 - 2020-09-21